Create adapter pack

+++ app/src/main/java/com/example/shoppinglist/domain/ShopItem.kt
@@ -1,6 +1,6 @@
 package com.example.shoppinglist.domain
 
-data class ShopItem(val name: String, val count: Float,val enabled: Boolean,var id:Int = UNDEFINED_ID)
+data class ShopItem(val name: String, val count: Int, val enabled: Boolean, var id:Int = UNDEFINED_ID)
 {
 
     companion object{
